dwm: removal of pertag and notitle patches
Removes `pertag` patch after quickly testing it in real-life usage for a
day. Behaves more like herbstluftwm etc, but now that I'm back to dwm
and the good old dynamic tiling behaviour, this patch felt quickly
counter-intuitive. It feels much better to keep the same layout in my
mental view and to bring on and off the desired tagged apps (windows) to
the current view as needed. Quite like described in the Lubutu's dwm FAQ
"Why don't tags remember their layout?"
<http://web.archive.org/web/20120120161931/http://lubutu.com/rant/dwm-faq>
One screencast where this is nicely demonstrated:
<https://www.youtube.com/watch?v=GQ5s6T25jCc>

Removes `notitle` patch after quickly testing it in real-life usage for
a day. Having title always displayed is perhaps not necessary and a bit
less zen to the screen real estate, but I felt a need for it in several
cases, for example to see the page numbers when viewing PDF documents
with `mupdf`.

Signed-off-by: Tibor Šimko <tibor.simko@cern.ch>