dwm: removes hide_vacant_tags patch
Removes hide_vacant_tags patch and returns to the stock full-tag list
display in the status bar.

The main motivation is, when viewing all windows via "Mod-0", it is
sometimes useful to see which tags are "occupied" and which are "empty".
The stock tag box allows to see this by means of extra tag indicators in
the superscript. The hide_vacant_tags patch does not allow to see this
in the desired styling.

Furthermore, dwm's active/occupied tag indicators are aesthetically
alike to tmux's active/previous/etc window indicators, creating nicer
visual harmony of the whole dwm/tmux experience.

Signed-off-by: Tibor Šimko <tibor.simko@cern.ch>