dwm: push_no_master patch instead of movestack
Replaces `movestack` patch by the `push` patch that provides similar
functionality. However we apply the `push` patch in its `push_no_master`
variant, which allows to reorder clients only on the slave stack area.
I.e. we don't allow to push clients to the master area when reordering.
This is because we already have the "zoom" feature for that. This makes
for a nice separation of concerns when "pushing" clients (Mod-Shift-n/p)
vs "zooming" clients (Mod-Ret). The clients do not accidentally overflow
to the master area when reordering.

Signed-off-by: Tibor Šimko <tibor.simko@cern.ch>