tmux: no conditional hostname display
Removes conditional display of non-home hostname. It is not necessary,
since the tmux session context is usually known from its name. Moreover,
the non-home host name is usually displayed in the command-line prompt
and in the window title.  Hence it is not necessary to clutter tmux
status bar as well.

Signed-off-by: Tibor Šimko <tibor.simko@cern.ch>