dwm: focusonnetactive patch
Switches to the `focusonnetactive` patch instead of using the
`focusurgent` patch. This change automatically raises the concerned
window instead of just setting the urgency flag to which one has to hop
on via a key press. Allows to open URLs and have browser window focused
automatically. Allows to jump to wmctrl-dmenu-selected window
automatically.

Signed-off-by: Tibor Šimko <tibor.simko@cern.ch>