emacs: garbage collection to happen only when idle
Configures garbage collection to happen only when Emacs is idle,
thanks to the gcmh package.

This leads to swifter runtime performance during regular operation, as
the garbage collection does not hit in the middle of heavy completions
and other operations.

Keeping default garbage collection threshold of about 1 GB for now.

Turned the verbosity on to observe garbage collection performance for
now.