global: move from xterm to alacritty
Moving from xterm to alacritty, for several reasons.

Firstly, when using the EXWM window manager, it is not possible to
send ^C to xterm terminals unless allowSendEvents resource is set.
However, setting this resource inactivates other allow*Ops resources
due to security concerns, which means that the window title cannot be
set dynamically to the current path. Alacritty does not have this
problem.

Secondly, Alacritty has a nice vi-like copy mode similar to Termite,
very nice to copy-paste terminal output without mouse. The author of
Termite actually stopped development and now recommends to switch to
Alacritty.

(Last, the speed of Alactritty is just a nice side effect, not the
principal reason for the switch.)