emacs: fix vterm to be vi-mode friendly
Fixes vterm so that, OT1H, the standard Emacs mode is used in prompt
instead of Evil mode, and the prompt then relies on zsh shell being
configured in vi-mode. (This is necessary because vterm is not
evil-mode-friendly out of the box.) OTOH, it keeps evil-mode in the
copy mode which is useful when copy-pasting terminal output.  Inspired
by @meliache.