starship: initial prompt configuration
Adds initial configuration of Starship prompt. Uses textual instead of
graphical symbols for languages and whatnot. Disables most modules for
speed. Configures look and feel to be similar to Pure prompt.

One easy advantage of Starship over Pure prompt is that the
out-of-the-box behaviour of the second-line prompt continuation in
Starship is to use vanilla dollar prompt symbol only, which makes it
very easy to copy/paste terminal command examples to chat or docs.
(Pure uses venv on the second-line prompt continuation.)