xterm: set as preferred terminal emulator
Switches back to xterm as the preferred terminal emulator. The reason
being that I compared alacritty, kitty and xterm for the same buffer
and the same size, and on my particular setup of preferred font and
preferred monitor resolution, the xterm display was the sharpest of
the three. This was especially noticeable on letters such as
`w`. Confirmed by an independent pair of eyes.