vim: removes fzf ev/eh shortcuts
Removes shortcuts to run fzf in vertical and horizontal splits that I
rarely used. It is sufficient to use the main fzf functionality and then
use `C-v` instead of `RET` to open the selection in a new vertical
split, `C-x` in a new horizontal split, and `C-t` in a new tab.