vim: fzf lines shortcut to use current word
When invoking fzf lines shortcut, use current word to as a preselection.
This is usually the operation most often wanted, and if the word is
wrong, then it is easy to hit `C-w` to remove it.

Also, changes the shortcut to mnemotechnic `<leader>l`, since this
key combination has now been freed.