vim: initial CoC configuration for LSP
Adds initial CoC confguration for LSP functionality. Coc seems to
perform better than vim-lsp out of the box. For example, when using `K`
to obtain a documentation with pyright in a popup window, with vim-lsp
the arrow keys started to garble the terminal output and produced ABCD
letters on the terminal. CoC does not have this problem and works nicely
out of the box and displays also type hints from pyright out of the box.