awesome: fix xterm tiled window size
AwesomeWM respects window size hits by default, so that e.g. xterm
windows were created with perfect size from the number-of-characters
point of view (depending on the available space but also on the font
size). This was however making tiled windows "unequally sized" from the
number-of-pixels point of view when several xterm clients appeared on
the same tag or when an xterm client appeared next to browser on the
same tag.

This commit asks AwesomeWM to disregard client size hits, making all
tiled windows nicely equal in size. The "extra" space where xterm cannot
print whole characters at the given font size anymore is simply treated
as a sort of "internal padding".