tmux: simpler status line
Simplifies Tmux status line, leaving out the battery and day
information, now that the AwesomeWM bar is more compact and includes
also the battery status. I'm usually having the WM bar "on" anyway, as
it is useful for viewing several tags at once, or for identifying
currently focused client thanks to the WM client task list.

This commit therefore leaves only the time information in the Tmux
status line, which is still useful to have when I use the terminal
fullscreen in order to focus. And, when I need to consult the battery
status, I can quickly "unfullscreen" the terminal via `s-f` to check the
WM status bar.